backend/requirements.txt
anyio
autopep8
certifi
charset-normalizer
clic
colorama
dnspython
email-validator
fastapi
fastapi-utils
greenlet
h11
httptools
idna
itsdangerous
Jinja2
MarkupSafe
orjson
pycodestyle
pydantic
python-dotenv
python-multipart
PyYAML
requests
six
sniffio
SQLAlchemy
sqlalchemy-migrate
starlette
tomli
typing_extensions
ujson
urllib3
uvicorn
watchfiles
websockets
mysqlclient
mysql-connector
python-slugify
jsons
faker
pymysql
passlib[bcrypt]
python-jose[cryptography]
cbfa
password-strength
jinja2
importmonkey
PyJWT
bcrypt==3.2.2
jsonpickle